requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.7
Django==4.2.3
django-autoslug==1.9.9
django-filter==23.2
django-jet-reboot==1.3.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.12.2
flake8==6.0.0
identify==2.5.26
inflection==0.5.1
isort==5.12.0
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.1
phonenumbers==8.13.17
Pillow==9.5.0
platformdirs==3.9.1
pre-commit==3.3.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
virtualenv==20.24.2
